+++ src/main/java/org/folio/services/item/ItemService.java
@@ -1,14 +1,8 @@
 package org.folio.services.item;
 
-import static io.vertx.core.CompositeFuture.all;
-import static io.vertx.core.Future.failedFuture;
 import static io.vertx.core.Future.succeededFuture;
 import static io.vertx.core.Promise.promise;
-import static java.lang.String.format;
-import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.toList;
-import static org.apache.commons.lang3.StringUtils.isBlank;
-import static org.apache.logging.log4j.LogManager.getLogger;
 import static org.folio.rest.impl.ItemStorageAPI.ITEM_TABLE;
 import static org.folio.rest.impl.StorageHelper.MAX_ENTITIES;
 import static org.folio.rest.jaxrs.resource.ItemStorage.DeleteItemStorageItemsByItemIdResponse;
@@ -21,28 +15,8 @@ import static org.folio.rest.persist.PgUtil.postSync;
 import static org.folio.rest.persist.PgUtil.postgresClient;
 import static org.folio.rest.persist.PgUtil.put;
 import static org.folio.rest.support.CollectionUtil.deepCopy;
-
-import java.util.Collection;
-import java.util.Date;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.function.Function;
-
-import javax.ws.rs.core.Response;
-
-import org.apache.logging.log4j.Logger;
-import org.folio.persist.ItemRepository;
-import org.folio.rest.exceptions.BadRequestException;
-import org.folio.rest.jaxrs.model.HoldingsRecord;
-import org.folio.rest.jaxrs.model.Item;
-import org.folio.rest.persist.PostgresClient;
-import org.folio.rest.persist.SQLConnection;
-import org.folio.rest.support.HridManager;
-import org.folio.services.ItemEffectiveValuesService;
-import org.folio.services.batch.BatchOperationContext;
-import org.folio.services.domainevent.ItemDomainEventPublisher;
-import org.folio.validator.CommonValidators;
+import static org.folio.services.batch.BatchOperationContextFactory.buildBatchOperationContext;
+import static org.folio.validator.HridValidators.refuseWhenHridChanged;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
@@ -50,10 +24,23 @@ import io.vertx.core.Future;
 import io.vertx.core.Promise;
 import io.vertx.sqlclient.Row;
 import io.vertx.sqlclient.RowSet;
+import java.util.Collection;
+import java.util.Date;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import javax.ws.rs.core.Response;
+import org.folio.persist.ItemRepository;
+import org.folio.rest.jaxrs.model.HoldingsRecord;
+import org.folio.rest.jaxrs.model.Item;
+import org.folio.rest.persist.PostgresClient;
+import org.folio.rest.persist.SQLConnection;
+import org.folio.rest.support.HridManager;
+import org.folio.services.ItemEffectiveValuesService;
+import org.folio.services.domainevent.ItemDomainEventPublisher;
+import org.folio.validator.CommonValidators;
 
 public class ItemService {
-  private static final Logger log = getLogger(ItemService.class);
-
   private final HridManager hridManager;
   private final ItemEffectiveValuesService effectiveValuesService;
   private final Context vertxContext;
@@ -75,7 +62,7 @@ public class ItemService {
   public Future<Response> createItem(Item entity) {
     entity.getStatus().setDate(new Date());
 
-    return setHrid(entity)
+    return hridManager.populateHrid(entity)
       .compose(effectiveValuesService::populateEffectiveValues)
       .compose(item -> {
         final Promise<Response> postResponse = promise();
@@ -90,19 +77,13 @@ public class ItemService {
 
   public Future<Response> createItems(List<Item> items, boolean upsert) {
     final Date itemStatusDate = new java.util.Date();
+    items.stream()
+      .filter(item -> item.getStatus().getDate() == null)
+      .forEach(item -> item.getStatus().setDate(itemStatusDate));
 
-    @SuppressWarnings("all")
-    final List<Future> setHridFutures = items.stream()
-      .map(item -> {
-        if (item.getStatus().getDate() == null) {
-          item.getStatus().setDate(itemStatusDate);
-        }
-        return setHrid(item);
-      }).collect(toList());
-
-    return all(setHridFutures)
+    return hridManager.populateHridForItems(items)
       .compose(result -> effectiveValuesService.populateEffectiveValues(items))
-      .compose(result -> buildBatchOperationContext(upsert, items))
+      .compose(result -> buildBatchOperationContext(upsert, items, itemRepository, Item::getId))
       .compose(batchOperation -> {
         final Promise<Response> postSyncResult = promise();
 
@@ -188,39 +169,5 @@ public class ItemService {
 
   private Function<SQLConnection, Future<RowSet<Row>>> updateSingleItemBatchFactory(Item item) {
     return connection -> itemRepository.update(connection, item.getId(), item);
-  }
-
-  private Future<Item> refuseWhenHridChanged(Item oldItem, Item newItem) {
-    if (Objects.equals(newItem.getHrid(), oldItem.getHrid())) {
-      return succeededFuture(oldItem);
-    } else {
-      log.warn("HRID has been changed for item {}", oldItem.getId());
-      return failedFuture(new BadRequestException(format(
-        "The hrid field cannot be changed: new=%s, old=%s", newItem.getHrid(),
-        oldItem.getHrid())));
-    }
-  }
-
-  private Future<Item> setHrid(Item entity) {
-    return isBlank(entity.getHrid())
-      ? hridManager.getNextItemHrid().map(entity::withHrid)
-      : succeededFuture(entity);
-  }
-
-  private Future<BatchOperationContext<Item>> buildBatchOperationContext(
-    boolean upsert, List<Item> allItems) {
-
-    if (!upsert) {
-      return succeededFuture(new BatchOperationContext<>(allItems, emptyList()));
-    }
-
-    return itemRepository.getById(allItems, Item::getId)
-      .map(foundItems -> {
-        final var itemsToBeCreated = allItems.stream()
-          .filter(item -> !foundItems.containsKey(item.getId()))
-          .collect(toList());
-
-        return new BatchOperationContext<>(itemsToBeCreated, foundItems.values());
-      });
   }
 }