+++ src/test/java/org/folio/rest/api/InstanceStorageTest.java
@@ -15,6 +15,8 @@ import org.folio.rest.jaxrs.model.Errors;
 import org.folio.rest.jaxrs.model.InstancesBatchResponse;
 import org.folio.rest.jaxrs.model.MarcJson;
 import org.folio.rest.jaxrs.model.NatureOfContentTerm;
+import org.folio.rest.jaxrs.model.Publication;
+import org.folio.rest.jaxrs.model.PublicationPeriod;
 import org.folio.rest.persist.PgUtil;
 import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.support.*;
@@ -64,9 +66,6 @@ import static org.hamcrest.CoreMatchers.hasItem;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.both;
-import static org.hamcrest.Matchers.either;
-import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.*;
 import static org.hamcrest.collection.IsIterableContainingInAnyOrder.containsInAnyOrder;
 import static org.hamcrest.core.IsNull.notNullValue;
@@ -129,9 +128,11 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
       .map(NatureOfContentTerm::getId)
       .toArray(String[]::new);
 
+    var publicationPeriod = new PublicationPeriod().withStart(2000).withEnd(2001);
+
     JsonObject instanceToCreate = smallAngryPlanet(id);
-    instanceToCreate.put("natureOfContentTermIds", Arrays
-      .asList(natureOfContentIds));
+    instanceToCreate.put("natureOfContentTermIds", Arrays.asList(natureOfContentIds));
+    instanceToCreate.put("publicationPeriod", JsonObject.mapFrom(publicationPeriod));
 
     CompletableFuture<Response> createCompleted = new CompletableFuture<>();
 
@@ -180,6 +181,11 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     assertThat(instanceFromGet.getBoolean(DISCOVERY_SUPPRESS), is(false));
     assertCreateEventForInstance(instanceFromGet);
+
+    var storedPublicationPeriod = instance.getJsonObject("publicationPeriod")
+      .mapTo(PublicationPeriod.class);
+    assertThat(storedPublicationPeriod.getStart(), is(2000));
+    assertThat(storedPublicationPeriod.getEnd(), is(2001));
   }
 
   @Test
@@ -493,7 +499,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     InterruptedException,
     ExecutionException,
     TimeoutException {
-    
+
     createInstancesWithClassificationNumbers();
 
     JsonObject allInstances = searchForInstances("classifications =/@classificationNumber \"K1 .M385\"");
@@ -508,7 +514,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     InterruptedException,
     ExecutionException,
     TimeoutException {
-    
+
     createInstancesWithClassificationNumbers();
 
     JsonObject allInstances = searchForInstances("classifications =\"K1 .M385\"");
@@ -2129,8 +2135,8 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     assertThat(response.getStatusCode(), is(400));
     assertThat(response.getBody(),
-        is("duplicate key value violates unique constraint \"instance_hrid_idx_unique\": " +
-          "Key (lower(f_unaccent(jsonb ->> 'hrid'::text)))=(in00000000001) already exists."));
+        is("lower(f_unaccent(jsonb ->> 'hrid'::text)) value already " +
+          "exists in table instance: in00000000001"));
 
     log.info("Finished cannotCreateInstanceWithDuplicateHRID");
   }
@@ -2580,8 +2586,8 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     assertThat(response.getStatusCode(), is(400));
     assertThat(response.getBody(),
-        is("duplicate key value violates unique constraint \"instance_matchkey_idx_unique\": " +
-          "Key (lower(f_unaccent(jsonb ->> 'matchKey'::text)))=(match_key) already exists."));
+        is("lower(f_unaccent(jsonb ->> 'matchKey'::text)) value already " +
+          "exists in table instance: match_key"));
 
     log.info("Finished cannotCreateInstanceWithDuplicateMatchKey");
   }